services/orders/src/events/subscribers/ticketCreatedSubscriber.ts
import { Subjects, Subscriber, TicketCreatedEvent } from "@trensetickets/packages/pubsub";
import { Message } from "node-nats-streaming";
import { events } from "../../constants";
import { Ticket } from "../../models/ticket";

export class TicketCreatedSubscriber extends Subscriber<TicketCreatedEvent> {
  readonly subject = Subjects.TicketCreated;
  queueGroupName = events.queueGroupName;

  async onMessage(data: TicketCreatedEvent["data"], msg: Message) {
    const { id, title, price } = data;

    const ticket = Ticket.build({ id, title, price });
    await ticket.save();

    msg.ack();
  }
}
